Update requirement, Jenkins and pipeline temlpates.
