+++ auth-api/jenkins/prod.groovy
@@ -33,6 +33,17 @@ def NAMESPACE_DEPLOY = "${NAMESPACE_APP}" + '-' + "${DESTINATION_TAG}"
 
 def ROCKETCHAT_DEVELOPER_CHANNEL='#relationship-developers'
 
+// Get an image's hash tag
+String getImageTagHash(String imageName, String tag = "") {
+
+    if(!tag?.trim()) {
+        tag = "latest"
+    }
+
+    def istag = openshift.raw("get istag ${imageName}:${tag} -o template --template='{{.image.dockerImageReference}}'")
+    return istag.out.tokenize('@')[1].trim()
+}
+
 // post a notification to rocketchat
 def rocketChatNotificaiton(token, channel, comments) {
   def payload = JsonOutput.toJson([text: comments, channel: channel])
@@ -58,10 +69,10 @@ node {
                 }
                 openshift.withCluster() {
                     openshift.withProject("${NAMESPACE_BUILD}") {
-                        // echo "Tagging ${APP_NAME}:${DESTINATION_TAG}-prev ..."
-                        // def IMAGE_HASH = getImageTagHash("${APP_NAME}", "${DESTINATION_TAG}")
-                        // echo "IMAGE_HASH: ${IMAGE_HASH}"
-                        // openshift.tag("${APP_NAME}@${IMAGE_HASH}", "${APP_NAME}:${DESTINATION_TAG}-prev")
+                        echo "Tagging ${APP_NAME}:${DESTINATION_TAG}-prev ..."
+                        def IMAGE_HASH = getImageTagHash("${APP_NAME}", "${DESTINATION_TAG}")
+                        echo "IMAGE_HASH: ${IMAGE_HASH}"
+                        openshift.tag("${APP_NAME}@${IMAGE_HASH}", "${APP_NAME}:${DESTINATION_TAG}-prev")
 
                         echo "Tagging ${APP_NAME} for deployment to ${DESTINATION_TAG} ..."
 						openshift.tag("${APP_NAME}:${SOURCE_TAG}", "${APP_NAME}:${DESTINATION_TAG}")
