+++ openshift/templates/api-config.json
@@ -48,7 +48,15 @@
         "TOKEN_EXPIRY_PERIOD": "${TOKEN_EXPIRY_PERIOD}",
         "LEGAL_API_URL": "",
         "SERVICE_SCHEDULE": "",
-        "NOTIFY_API_URL": ""
+        "NOTIFY_API_URL": "",
+        "NATS_CLIENT_NAME": "",
+        "NATS_CLUSTER_ID": "",
+        "NATS_QUEUE": "",
+        "NATS_SERVERS": "",
+        "NATS_SUBJECT": "",
+        "DELIVERY_FAILURE_RETRY_TIME_FRAME": "",
+        "FAILURE_EMAIL_TIME_FRAME": "",
+        "PENDING_EMAIL_TIME_FRAME": ""
       }
     }
   ],
