+++ queue_services/notify-queue/jenkins/prod.groovy
@@ -22,7 +22,7 @@
 import groovy.json.*
 
 // define constants - values sent in as env vars from whatever calls this pipeline
-def APP_NAME = 'notify-api'
+def APP_NAME = 'notify-queue'
 def SOURCE_TAG = 'test'
 def DESTINATION_TAG = 'prod'
 def TOOLS_TAG = 'tools'
@@ -32,6 +32,17 @@ def NAMESPACE_BUILD = "${NAMESPACE_APP}"  + '-' + "${TOOLS_TAG}"
 def NAMESPACE_DEPLOY = "${NAMESPACE_APP}" + '-' + "${DESTINATION_TAG}"
 
 def ROCKETCHAT_DEVELOPER_CHANNEL='#relationship-developers'
+
+// Get an image's hash tag
+String getImageTagHash(String imageName, String tag = "") {
+
+    if(!tag?.trim()) {
+        tag = "latest"
+    }
+
+    def istag = openshift.raw("get istag ${imageName}:${tag} -o template --template='{{.image.dockerImageReference}}'")
+    return istag.out.tokenize('@')[1].trim()
+}
 
 // post a notify to rocketchat
 def rocketChatNotificaiton(token, channel, comments) {
@@ -58,6 +69,11 @@ node {
                 }
                 openshift.withCluster() {
                     openshift.withProject("${NAMESPACE_BUILD}") {
+                        echo "Tagging ${APP_NAME}:${DESTINATION_TAG}-prev ..."
+                        def IMAGE_HASH = getImageTagHash("${APP_NAME}", "${DESTINATION_TAG}")
+                        echo "IMAGE_HASH: ${IMAGE_HASH}"
+                        openshift.tag("${APP_NAME}@${IMAGE_HASH}", "${APP_NAME}:${DESTINATION_TAG}-prev")
+
                         echo "Tagging ${APP_NAME} for deployment to ${DESTINATION_TAG} ..."
 						openshift.tag("${APP_NAME}:${SOURCE_TAG}", "${APP_NAME}:${DESTINATION_TAG}")
                     }
