+++ queue_services/notify-queue/jenkins/test.groovy
@@ -22,7 +22,7 @@
 import groovy.json.*
 
 // define constants - values sent in as env vars from whatever calls this pipeline
-def APP_NAME = 'notify-api'
+def APP_NAME = 'notify-queue'
 def SOURCE_TAG = 'dev'
 def DESTINATION_TAG = 'test'
 def TOOLS_TAG = 'tools'
