+++ queue_services/notify-queue/requirements/repo-libraries.txt
@@ -1,2 +1,2 @@
-git+https://github.com/bcgov/sbc-auth.git@development#egg=notify_api&subdirectory=notify_api
+git+https://github.com/bcgov/sbc-auth.git@development#egg=notify_api&subdirectory=notify-api
 git+https://github.com/bcgov/lear.git#egg=entity_queue_common&subdirectory=queue_services/common
